Moved RecentChips into RecentChipsContainer folder
