+++ src/components/ProfileDialog/index.ts
@@ -1,0 +1,1 @@
+export { ProfileDialog } from "./ProfileDialog";
