+++ src/components/RecentChipsContainer/index.ts
@@ -1,0 +1,1 @@
+export { RecentChipsContainer } from "./RecentChipsContainer";
